travel/requirements.txt
Flask==1.1.1
mysql-connector==2.2.9
Flask-SQLAlchemy==1.0
mysqlclient==1.4.1
Flask-MySQLdb==0.2.0
gunicorn
requests

